src/requirements.txt
Flask
google-cloud-firestore
google-cloud-logging
gunicorn
jsonschema
